+++ backend-javaspring/src/main/java/org/example/app/services/PublisherService.java
@@ -2,11 +2,10 @@ package org.example.app.services;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.example.app.models.entities.UserEntity;
 import org.example.app.models.requests.PublisherRequestDTO;
 import org.example.app.models.entities.PublisherEntity;
 import org.example.app.models.responses.PublisherResponseDTO;
-import org.example.app.models.responses.UserResponseDTO;
+import org.example.app.repositories.BookRepository;
 import org.example.app.repositories.PublisherRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -19,6 +18,7 @@ import java.util.Optional;
 public class PublisherService {
 
     private final PublisherRepository publisherRepository;
+    private final BookRepository bookRepository;
     private final ModelMapper modelMapper;
 
     @Transactional
@@ -83,6 +83,10 @@ public class PublisherService {
     public void deleteService(Long id) {
         if (!publisherRepository.existsById(id)) {
             throw new RuntimeException("Editora com o id " + id + " não encontrado.");
+        }
+
+        if (bookRepository.existsByPublisher_Id(id)) {
+            throw new RuntimeException("Editora com o id " + id + " está vinculada a um livro e não pode ser deletada.");
         }
 
         publisherRepository.deleteById(id);
